Move navigation bar to a base activity
