+++ app/src/main/res/layout/app_bar_daily_log.xml
@@ -20,15 +20,4 @@
 
     </com.google.android.material.appbar.AppBarLayout>
 
-    <include layout="@layout/content_daily_log" />
-
-
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/fab"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_margin="@dimen/fab_margin"
-        app:srcCompat="@android:drawable/ic_dialog_email" />
-
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
